+++ src/SelbaWard/Starfield.cpp
@@ -133,7 +133,7 @@ void Starfield::setColor(const sf::Color color)
 void Starfield::draw(sf::RenderTarget& target, sf::RenderStates states) const
 {
 	states.transform *= getTransform();
-	const unsigned int size{ m_vertices.size() };
+	const unsigned int size{ static_cast<unsigned int>(m_vertices.size()) };
 	if (size > 0)
 		target.draw(&m_vertices.front(), size, m_primitiveType, states);
 }
